Fix profile saver hotness for AOT code
If the counter is larger than the warm threshold, mark the method as
hot. This fixes a bug for profiling arleady compiled methods where
they would not be marked as hot no matter what the counter was.

Also removed unused functions, and reverted HasSeenMethod to verify
that the method is hot.

Bug: 139883463
Test: test-art-host

Actual commit" Revert^2 "Fix profile saver hotness for AOT code"

Change-Id: I243e528e9f06da7223eed9b91ad613d4f1ad9524