ARM64: Change code emitted by ClinitCheck.
Change the code from MVN+CBNZ to CMP+BLO. The latter is
better optimized in ARM64 CPUs. To avoid increasing code
size, this requires the preceding load to be changed from
LDR to LDRB for a single byte of the 32-bit field.

This shows small but measurable improvement on a few Golem
benchmarks, for example MicroLambda, KotlinAutoReversiBench
and KotlinImgProc-GaussianBlurOpt.

Test: testrunner.py --target --optimizing
Bug: 36692143
Change-Id: Ia73f791d7026220ef38e73bd5ee19fcc4877564d
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>